Corrected Categories v1 Categories cs renamed for MenuCategories cs because there was a conflict between our module and Hotcakes module.
