+++ Controllers/Class1.cs
@@ -1,0 +1,14 @@
+﻿using Microsoft.Analytics.Interfaces;
+using Microsoft.Analytics.Types.Sql;
+using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using System.Text;
+
+namespace Company.Modules.MenuModul.Controllers
+{
+    internal class Class1
+    {
+    }
+}
