+++ Controllers/TestController.cs
@@ -1,0 +1,24 @@
+﻿using DotNetNuke.Collections;
+using DotNetNuke.Security;
+using DotNetNuke.Web.Mvc.Framework.ActionFilters;
+using DotNetNuke.Web.Mvc.Framework.Controllers;
+using System.Web.Mvc;
+namespace MvcMovie.Controllers
+{
+
+    public class TestController : DnnController
+    {
+        // 
+        // GET: /HelloWorld/
+        public string Index()
+        {
+            return "This is my default action...";
+        }
+        // 
+        // GET: /HelloWorld/Welcome/ 
+        public string Welcome()
+        {
+            return "This is the Welcome action method...";
+        }
+    }
+}
