+++ Models/Test.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using System.Text;
+
+namespace Company.Modules.MenuModul.Models
+{
+    internal class Test
+    {
+    }
+}
