+++ Views/Item/Edit.cshtml
@@ -1,0 +1,51 @@
+﻿@inherits DotNetNuke.Web.Mvc.Framework.DnnWebViewPage<Company.Modules.MenuModul.Models.Item>
+
+    ”
+    @using System.Collections.Generic
+    @using DotNetNuke.Web.Mvc.Helpers
+
+    <div id="Item-@Dnn.ModuleContext.ModuleId">
+        <div class="dnnForm dnnEditBasicSettings" id="dnnEditBasicSettings">
+            <div class="dnnFormExpandContent dnnRight "><a href="#">@Dnn.LocalizeString("ExpandAll")</a></div>
+
+            <h2 id="dnnSitePanel-BasicSettings" class="dnnFormSectionHead dnnClear">
+                <a href="#" class="dnnSectionExpanded">@Dnn.LocalizeString("BasicSettings")</a>
+            </h2>
+            <fieldset>
+                <div class="dnnFormItem">
+                    <div><label for="itemName">@Dnn.LocalizeString("lblName")</label></div>
+                    @Html.TextBoxFor(m => m.ItemName)
+                </div>
+                <div class="dnnFormItem">
+                    <div><label for="itemDescription">@Dnn.LocalizeString("lblDescription")</label></div>
+                    @Html.TextArea("itemDescription", Model.ItemDescription)
+                </div>
+                <div class="dnnFormItem">
+                    <div><label for="itemUser">@Dnn.LocalizeString("lblAssignedUser")</label></div>
+                    @Html.DropDownListFor(m => m.AssignedUserId, (IEnumerable<SelectListItem>)ViewBag.Users, "Choose...")
+                </div>
+                @Html.HiddenFor(m => m.ModuleId)
+            </fieldset>
+        </div>
+        <button type="submit" class="dnnPrimaryAction">@Dnn.LocalizeString("Submit")</button>
+        <a id="cancelEdit" href="#" class="dnnSecondaryAction">@Dnn.LocalizeString("Cancel")</a>
+    </div>
+
+    <script type="text/javascript">
+    /*globals jQuery, window, Sys */
+    (function($, Sys) {
+        function dnnEditBasicSettings() {
+            $('#dnnEditBasicSettings').dnnPanels();
+            $('#dnnEditBasicSettings .dnnFormExpandContent a').dnnExpandAll({ expandText: '@Dnn.LocalizeString("ExpandAll")', collapseText: '@Dnn.LocalizeString("CollapseAll")', targetArea: '#dnnEditBasicSettings' });
+        }
+
+        $(function() {
+            $('#cancelEdit').click(function() { dnnModal.closePopUp(false); });
+            dnnEditBasicSettings();
+            Sys.WebForms.PageRequestManager.getInstance().add_endRequest(function() {
+                dnnEditBasicSettings();
+            });
+        });
+
+    }(jQuery, window.Sys));
+    </script>
