+++ Views/Test/Test.cshtml
@@ -1,0 +1,8 @@
+﻿@using DotNetNuke.Web.Mvc.Helpers
+
+<h2>Index</h2>
+
+<p>Hello from our View Template!</p>
+
+
+
